requirements.txt
certifi==2017.11.5
chardet==3.0.4
dj-stripe==1.0.0.post1
Django==1.11.7
django-appconf==1.0.2
django-cors-headers==2.1.0
django-imagekit==4.0.1
django-polymorphic==1.3
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
idna==2.6
jsonfield==2.0.2
mysqlclient==1.3.12
olefile==0.44
pilkit==2.0
Pillow==4.3.0
PyJWT==1.5.3
python-doc-inherit==0.3.0
pytz==2017.3
requests==2.18.4
six==1.11.0
stripe==1.75.0
tqdm==4.19.4
urllib3==1.22
